Code cleanup: Separate utilities
