--- src/utils/portals-common/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "portals-common",
-  "version": "0.0.0",
-  "private": true,
-  "main": "./portals-common.js"
-}
