+++ src/utils/station-filtering/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "station-filtering",
+  "version": "0.0.0",
+  "private": true,
+  "main": "station-filtering.js"
+}
