+++ src/utils/station-filtering/station-filtering.js
@@ -4,85 +4,18 @@ import filter from 'lodash/fp/filter';
 import flatten from 'lodash/fp/flatten';
 import flow from 'lodash/fp/flow';
 import map from 'lodash/fp/map';
-import min from 'lodash/fp/min';
-import max from 'lodash/fp/max';
 import some from 'lodash/fp/some';
 import uniq from 'lodash/fp/uniq';
 import isNil from 'lodash/fp/isNil';
-import flattenDepth from 'lodash/fp/flattenDepth';
-import { isPointInPolygonWn } from '../geometry-algorithms';
+import { utcDateOnly } from '../dates';
+import {
+  checkGeoJSONMultiPolygon, gJMultiPolygonBoundingBox, isPointInGeoJSONPolygon,
+} from '../geoJSON';
 import { stationVariableUris, uniqStationLocations } from '../station-info';
 import { getTimer } from '../timing';
 
 const ft = getTimer("Station filtering timing")
 ft.log();
-
-
-// GeoJSON MultiPolygon format. Taken from
-// https://conservancy.umn.edu/bitstream/handle/11299/210208/GeoJSON_Primer_2019.pdf
-//
-// {
-//   "type": "MultiPolygon",
-//   "coordinates": [
-//     // one or more Polygon coordinate array:
-//     [
-//       // one or more Linear ring coordinate arrays:
-//       [
-//          // at least four Points; first point = last point:
-//         [x0, y0],
-//         [x1, y1],
-//         [x2, y2],
-//         // ...
-//         [x0, y0]
-//       ],
-//       // ...
-//     ],
-//     // ...
-//   ],
-//   // ...
-// };
-
-
-const checkGeoJSONMultiPolygon = geometry => {
-  if (geometry['type'] !== 'MultiPolygon') {
-    throw new Error(`Invalid geometry type: ${geometry['type']}`)
-  }
-};
-
-
-const gJMultiPolygonBoundingBox = geometry => {
-  const points = flattenDepth(3, geometry["coordinates"]);
-  const xs = map(p => p[0], points);
-  const ys = map(p => p[1], points);
-  return [
-    [min(xs), max(ys)], // top left
-    [max(xs), min(ys)], // bottom right
-  ]
-};
-
-const getX = point => point[0];
-const getY = point => point[1];
-export const isPointInGeoJSONPolygon = isPointInPolygonWn(getX, getY);
-
-
-export const utcDateOnly = d => {
-  // Return a new Date object which is suitable for comparing UTC date only
-  // (fixed zero time component). One might suppose that the comparisons in
-  // the filter below also would work with local time (Date.setHours()), but
-  // they don't; only UTC works.
-  if (!d) {
-    return d;
-  }
-  const r = new Date(d);
-  r.setUTCHours(0, 0, 0, 0);
-  return r;
-};
-
-
-export const isNilOrAscending = (a, b, c) => {
-  // Allowing for nils (undefined, null), check that a <= b <= c.
-  return isNil(b) || ((isNil(a) || a <= b) && (isNil(c) || b <= c));
-};
 
 
 export const historyDateMatch = (
@@ -96,14 +29,10 @@ export const historyDateMatch = (
   endDate = utcDateOnly(endDate);
 
   if (strict) {
+    // TODO , or remove strict
     // Allowing for nils, check that
     // minObsDate <= startDate <= endDate <= maxObsDate.
-    // TODO: Optimize
-    return (
-      isNilOrAscending(startDate, endDate)
-      && isNilOrAscending(minObsDate, startDate, maxObsDate)
-      && isNilOrAscending(minObsDate, endDate, maxObsDate)
-    );
+    return true;
   }
 
   // This is a variant of the looser legacy comparison. Allows nil start, end.
